Rename pkg name
